refactor: move hooks to lib/hooks and remove unused hooks
- Deleted unused hooks: usePaginate, useApiQuery, useFragmentToken, useOrganization, useToast, useUserProfile.
- Moved existing hooks to lib/hooks directory for better organization.
- Updated imports across the application to reflect new hook locations.
- Added new implementations for useApiQuery, useFragmentToken, useOrganization, useToast, and useUserProfile in the lib/hooks directory.
- Ensured all hooks maintain their original functionality while improving code structure.
