+++ web/src/components/Auth.tsx
@@ -6,7 +6,7 @@ import { VStack } from '@astryxdesign/core/VStack';
 import { ApiError } from '@/lib/api';
 import NotFound from '@/platform/NotFound';
 import PlatformLayout from '@/platform/layout';
-import { useUserProfile } from '@/hooks/use-user';
+import { useUserProfile } from '@/lib/hooks/use-user';
 import { SignInCard } from '@/components/SignInCard';
 
 /** Protects routes and optionally restricts access to Platform administrators. */
