+++ web/src/lib/hooks/use-organization.tsx
@@ -6,8 +6,8 @@ import type {
     OrganizationMemberUpdate,
     OrganizationUpdate,
 } from '@/lib/generated/platform-api-v1/types.gen';
-import { useApiQuery } from '@/hooks/use-api';
-import { useUserProfile } from '@/hooks/use-user';
+import { useApiQuery } from '@/lib/hooks/use-api';
+import { useUserProfile } from '@/lib/hooks/use-user';
 import { platformApiPath } from '@/lib/platform-api';
 import { ApiError, apiQueryKey, fetchApiJson, requestApi } from '@/lib/api';
 import { zOrganizationDetails, zOrganizationSummary } from '@/lib/generated/platform-api-v1/zod.gen';
