+++ web/src/platform/root.tsx
@@ -3,7 +3,7 @@ import { useState, type ReactNode } from 'react';
 import { QueryClientProvider } from '@tanstack/react-query';
 import '@/index.css';
 import { AstryxProvider } from '@/providers';
-import { UserProvider } from '@/hooks/use-user';
+import { UserProvider } from '@/lib/hooks/use-user';
 import { createQueryClient } from '@/lib/react-query';
 import { Document } from '@/components/layouts/Document';
