+++ web/src/platform/routes/admin/Applications.tsx
@@ -11,9 +11,9 @@ import { Heading } from '@astryxdesign/core/Heading';
 import { EmptyState } from '@astryxdesign/core/EmptyState';
 import { pixel, proportional } from '@astryxdesign/core/Table';
 import type { ApplicationResponse, Status } from '@/lib/generated/platform-api-v1/types.gen';
-import { useApiQuery } from '@/hooks/use-api';
+import { useApiQuery } from '@/lib/hooks/use-api';
 import { dateTimeFormatter } from '@/lib/utils';
-import { usePaginate } from '@/hooks/pagination';
+import { usePaginate } from '@/lib/hooks/pagination';
 import { platformApiPath } from '@/lib/platform-api';
 import { Table, TableColumn } from '@/components/ui/Table';
 import { zApplicationResponse } from '@/lib/generated/platform-api-v1/zod.gen';
