+++ web/src/platform/routes/admin/Compute.tsx
@@ -10,10 +10,10 @@ import { pixel, proportional } from '@astryxdesign/core/Table';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { ComputeRegistryResponse } from '@/lib/generated/platform-api-v1/types.gen';
 import { requestApi } from '@/lib/api';
-import { useToast } from '@/hooks/use-toast';
-import { useApiQuery } from '@/hooks/use-api';
+import { useToast } from '@/lib/hooks/use-toast';
+import { useApiQuery } from '@/lib/hooks/use-api';
 import { useDeleteDialog } from '@/lib/utils';
-import { usePaginate } from '@/hooks/pagination';
+import { usePaginate } from '@/lib/hooks/pagination';
 import { computesQueryKey } from '@/lib/query-keys';
 import { platformApiPath } from '@/lib/platform-api';
 import { Table, TableColumn } from '@/components/ui/Table';
