+++ web/src/platform/routes/auth/Register.tsx
@@ -9,7 +9,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { TextInput } from '@astryxdesign/core/TextInput';
 import { Controller, useForm, useWatch } from 'react-hook-form';
 import { requestApiJson } from '@/lib/api';
-import { useToast } from '@/hooks/use-toast';
+import { useToast } from '@/lib/hooks/use-toast';
 import { AuthPage } from '@/components/AuthPage';
 import { platformApiPath } from '@/lib/platform-api';
 import { WelcomeTitle } from '@/components/WelcomeTitle';
